start of deletion of unnecessary files
the original repo prior to PNAS resubmission has been stored in a new branch. The main branch will now be tidied up and will contain only files relevant to the new version of the analysis on the expanded dataset along with some files relevant to supporting understanding of the maths and methods used to analyse the data.